msvc++ wants it different (2)
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@145 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ include/Options.h
@@ -71,7 +71,7 @@ typedef struct {
 	// Network
 	int		iNetworkPort;
 	int		iNetworkSpeed;
-	int		fUpdatePeriod;
+	float	fUpdatePeriod;
 
 	// Audio
 	int		iSoundOn;
